requirements.txt
python-telegram-bot>=21.6,<22
httpx>=0.27.0,<0.29
PyYAML>=6.0.2,<7
